backend/requirements.txt
black==25.1.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
coverage==7.8.2
dnspython==2.7.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
marshmallow==3.22.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mongoengine==0.29.1
mongomock==4.3.0
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pip-review==1.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pytest==8.4.0
pytest-cov==6.1.1
pytest-env==1.1.5
pytest-mock==3.14.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
RapidFuzz==3.10.1
requests==2.32.4
sentinels==1.0.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
